Exclude images with bad motion correction results from tomogram alignment (#122)
Relion's motion correction sometimes fails on local alignment and produces abstract images. These then get put into tomogram alignment and disrupt the outputs.

The only way I can readily see to avoid this is to check the star files written out by Relion.
Firstly, we need to check if the file exists, then if the local motion model was used, and finally when coefficients were found.
I've put a threshold of 1000 on the local alignment coefficients, but this is an arbitrary choice.

This will result in an increase in I/O and maybe slow down alignment slightly.