Tomogram improvements (#129)
A few changes to tomo alignment and reconstruction:

- Plot a scale bar for XY and XZ projection images
- Allow for post-reconstruction tomgram rotations to avoid having to use FlipVol in AreTomo which produces varying orientations
- Insert tomogram sizes into ispyb
- AlignTiltSeries jobs should take ExcludeTiltSeries as input
- Make apng images of unaligned stacks
- Increase the reconstruction volume to 1600 if a pre-tilt is given
- Provide a tilt axis to AreTomo, defaulting to 85

One question to consider is whether we want apng images of aligned stacks. That would probably required running AreTomo twice - once to get an aligned stack, then again to reconstuct. Or the reconstuction could be done with Relion if the pick locations don't match.