Added values used by Murfey's email notification workflow to the message sent back from cryolo (#136)
The PR at DiamondLightSource/python-murfey#484 looks for parameters named particle_count, astigmatism, defocus, and resolution in the messages received by the feedback queue, and these values appear to be calculated as part of the CTFFind service, but are not passed on back to Murfey.

This PR adds those values to the message sent by the Cryolo service back to Murfey.