Remove the dependency on the workflows frontend and common service (#145)
Ports in the parts of workflows Frontend and CommonService that we actually need.
This should allow us to simplify our RabbitMQ setup as we won't need the workflows exchanges any more.

The new code in this PR is fairly minimal, but there are a few repeated knock-on changes to the services and tests as every instance has needed modification.