Rationalise the cluster submission codes (#139)
The code used for slurm submission was excessively complicated, as there were two different files which duplicated methods.

This clarifies them into a service which submits wrappers, and a utility code that can be used by any service needing slurm restAPI capabilities.

Some changes to recipes are needed for this.