Create services for the classification jobs (#176)
Adds the option to run classification jobs as services.

These can take much longer than the rabbitmq timeout of 30 minutes, so rather than hold the message they acknowledge as soon as the process starts, then resend any failed messages to the queue.
To replicate the nacking behaviour on failure, they record the number of reinjections and nack after 5.

The diff for this PR is very large as the wrapper code had to be broken into separate functions, but the amount of new code is relatively small.