Service to pick particles with topaz (#132)
This will allow picking to be done using topaz rather than cryolo.
Useful for:
 -   Doppio-live
 -   Cases where someone has a trained topaz model
 -   Consensus picking between cryolo and topaz?

However, it isn't immediately usabe for the auto-processing as it requires a particle diameter to be input, and we currently have no way to get that in Murfey or to switch between cryolo and topaz.

This branch exists from a while ago when thinking about how to estimate the particle diameter automatically, but the service is functional without that so can go in.