Insert manual tilt offset into ispyb + naming fixes for cryolo and ctf output (#197)
If given a tilt offset, AreTomo2 does not calculate it and so it does not appear in the outputs. In that case we should insert the input value into ispyb.

Empirically this happens for angles >= 0.1

Also deals with a naming issue in cryolo, and finds alternative ctf output names for the case where MotionCor2 has been run.