Coordinate corrections for tomograms (#198)
Relion particle extraction failed to recognise the particle positions when un-centered in the tomogram. This applies the centering and unit conversion to the particles.star file made after picking.

Doing this requires more of the relion_options to be set. These should be set by the tomo_align service prior to picking, but they could be passed into picking as inputs if we wanted to be more sure of them being set.

These coordinate corrections appear to work for tomograms without a pre-tilt. The case of a pre-tilt is still to be investigated but these changes are useful enough to go in now anyway.