Tomogram re-run star file fixes (#199)
In the existing setup, if a tomogram is processed twice it will get inserted into star files twice. This resolves that and provides only a single insert.

For tomograms, denoising and segmentation this is as simple as not re-sending the node creator request.

For alignment parameters and picks we need to first remove the existing values from the star files, then add in the new values instead.