Do YZ projection rather than XZ if tilt axis is around zero (#217)
For display purposes we rotate tomograms differently depending on the tilt axis of the microscope. The existing projection setup assumed the tomogram being used was reconstructed with tilt axis around 90 degrees then rotated to match the tilts.

If the tilt axis is around 0 (e.g. eBIC Talos) this will result in projecting YZ which is heavily streaked. To avoid this we need to project along the other axis.